Crate bevy_ecs_tilemap

Bevy ECS Tilemap plugin is a ECS driven tilemap rendering library. It’s designed to be fast and highly customizable. Each tile is considered a unique entity and all tiles are stored in the game world.

Features

A tile per entity.

Fast rendering using a chunked approach.

Layers and sparse tile maps.

GPU powered animations.

Isometric and Hexagonal tile maps.

Initial support for Tiled file exports.

Support for isometric and hexagon rendering.

Built in animation support – see animation example.

Texture array support.

Modules

helpers

A module which provides helper functions.

map

A module which contains tilemap components.

prelude

A module which exports commonly used dependencies.

tiles

A module which contains tile components.

Structs

FrustumCulling

MaterialTilemapBundle

The default tilemap bundle. All of the components within are required.

TilemapFirstSet

TilemapPlugin

A bevy tilemap plugin. This must be included in order for everything to be rendered. But is not necessary if you are running without a renderer.

Type Aliases

TilemapBundle