bevy_ecs_tilemap::helpers::filling
Function fill_tilemap_hexagon

pub fn fill_tilemap_hexagon(
    texture_index: TileTextureIndex,
    origin: TilePos,
    radius: u32,
    hex_coord_system: HexCoordSystem,
    tilemap_id: TilemapId,
    commands: &mut Commands<'_, '_>,
    tile_storage: &mut TileStorage,
)

Fills a hexagonal region with the given tile_texture.

The rectangular region is defined by an origin in TilePos, and a radius.

Tiles that do not fit in the tilemap will not be created.