bevy_ecs_tilemap::helpers::square_grid
Module neighbors

Structs

Neighbors

Stores some data T associated with each neighboring grid cell, if present.

Enums

SquareDirection

The eight directions in which a neighbor may lie, on a square-like grid.

Constants

CARDINAL_SQUARE_DIRECTIONS

Array of cardinal `SquareDirections (N, W, S, E).

SQUARE_DIRECTIONS

Array of SquareDirection variants.

SQUARE_OFFSETS

Offsets of tiles that lie in each SquareDirection.