bevy_ecs_tilemap::helpers::square_grid::neighbors
Enum SquareDirection

pub enum SquareDirection {
    East,
    NorthEast,
    North,
    NorthWest,
    West,
    SouthWest,
    South,
    SouthEast,
}

The eight directions in which a neighbor may lie, on a square-like grid.

Note that isometric grids are also square-like grids. In particular, there is no difference between the grid system for square and diamond-isometric grids.

Variants

East

NorthEast

North

NorthWest

West

SouthWest

South

SouthEast

Implementations

impl SquareDirection

pub fn is_cardinal(&self) -> bool

Is this direction a cardinal direction (i.e. North, South, East, West)?

pub fn is_diagonal(&self) -> bool

Is this direction a diagonal direction (i.e. NorthEast, NorthWest, SouthWest, SouthEast)?