bevy_ecs_tilemap::helpers
Module transform

Functions

chunk_aabb

Calculates the Aabb of a generic chunk. The AABB depends upon the grid size, tile size, and map type of the the chunk the tile belongs to.

chunk_index_to_world_space

Calculates the world-space position of the bottom-left of the specified chunk.