bevy_ecs_tilemap
Module map

A module which contains tilemap components.

Structs

TilemapGridSize

Size of the tiles on the grid in pixels. This can be used to overlay tiles on top of each other. Ex. A 16x16 pixel tile can be overlapped by 8 pixels by using a grid size of 16x8.

TilemapId

A component which stores a reference to the tilemap entity.

TilemapRenderSettings

Custom parameters for the render pipeline.

TilemapSize

Size of the tilemap in tiles.

TilemapSpacing

Spacing between tiles in pixels inside of the texture atlas. Defaults to 0.0

TilemapTextureSize

Size of the atlas texture in pixels.

TilemapTileSize

Size of the tiles in pixels

Enums

HexCoordSystem

Different hex grid coordinate systems. You can find out more at this link: https://www.redblobgames.com/grids/hexagons/

IsoCoordSystem

Different isometric coordinate systems.

TilemapTexture

TilemapType

The type of tile to be rendered, currently we support: Square, Hex, and Isometric.

Constants

CHUNK_SIZE_2D

The default chunk_size (in tiles) used per mesh.