bevy_ecs_tilemap
Module prelude

A module which exports commonly used dependencies.

Re-exports

pub use crate::helpers;

pub use crate::MaterialTilemapBundle;

pub use crate::TilemapBundle;

pub use crate::TilemapPlugin;

pub use crate::helpers::filling::*;

pub use crate::helpers::geometry::*;

pub use crate::helpers::transform::*;

pub use crate::map::*;

pub use crate::tiles::*;

Structs

ArrayTextureLoader

A bevy world resource that allows you to add atlas textures for loading into the array texture cache.

MaterialTilemapHandle

MaterialTilemapKey

MaterialTilemapPlugin

StandardTilemapMaterial

TilemapArrayTexture

Traits

MaterialTilemap