Module bevy_ecs_tilemap::tiles

A module which contains tile components.

Structs

AnimatedTile

A component that is attached to a Tile entity that tells the GPU how to animate the tile. Currently all frames must be aligned in your tilemap.

TileBundle

This an optional tile bundle with default components.

TileColor

A custom color for the tile.

TileFlip

Flips the tiles texture along the X, Y or diagonal axes

TilePos

A tile position in the tilemap grid.

TilePosOld

TileStorage

Used to store tile entities for fast look up. Tile entities are stored in a grid. The grid is always filled with None.

TileTextureIndex

A texture index into the atlas or texture array for a single tile. Indices in an atlas are horizontal based.

TileVisible

Hides or shows a tile based on the boolean. Default: True