bevy_ecs_tilemap
Type Alias TilemapBundle

pub type TilemapBundle = MaterialTilemapBundle<StandardTilemapMaterial>;

Aliased Type

struct TilemapBundle {

    pub grid_size: TilemapGridSize,
    pub map_type: TilemapType,
    pub size: TilemapSize,
    pub spacing: TilemapSpacing,
    pub storage: TileStorage,
    pub texture: TilemapTexture,
    pub tile_size: TilemapTileSize,
    pub transform: Transform,
    pub global_transform: GlobalTransform,
    pub render_settings: TilemapRenderSettings,
    pub visibility: Visibility,
    pub inherited_visibility: InheritedVisibility,
    pub view_visibility: ViewVisibility,
    pub frustum_culling: FrustumCulling,
    pub material: MaterialTilemapHandle<StandardTilemapMaterial>,
    pub sync: SyncToRenderWorld,

}

Fields
grid_size: TilemapGridSizemap_type: TilemapTypesize: TilemapSizespacing: TilemapSpacingstorage: TileStoragetexture: TilemapTexturetile_size: TilemapTileSizetransform: Transformglobal_transform: GlobalTransformrender_settings: TilemapRenderSettingsvisibility: Visibility

User indication of whether an entity is visible

inherited_visibility: InheritedVisibility

Algorithmically-computed indication of whether an entity is visible and should be extracted for rendering

view_visibility: ViewVisibilityfrustum_culling: FrustumCulling

User indication of whether tilemap should be frustum culled.

material: MaterialTilemapHandle<StandardTilemapMaterial>sync: SyncToRenderWorld