Trait bevy_ecs_tilemap::prelude::MaterialTilemap

pub trait MaterialTilemap: AsBindGroup + Send + Sync + Clone + TypeUuid + TypePath + Sized + 'static {
    // Provided methods
    fn vertex_shader() -> ShaderRef { ... }
    fn fragment_shader() -> ShaderRef { ... }
    fn specialize(
        descriptor: &mut RenderPipelineDescriptor,
        key: MaterialTilemapKey<Self>
    ) { ... }
}

Provided Methods

fn vertex_shader() -> ShaderRef

Returns this material’s vertex shader. If ShaderRef::Default is returned, the default mesh vertex shader will be used.

fn fragment_shader() -> ShaderRef

Returns this material’s fragment shader. If ShaderRef::Default is returned, the default mesh fragment shader will be used.

fn specialize( descriptor: &mut RenderPipelineDescriptor, key: MaterialTilemapKey<Self> )

Customizes the default RenderPipelineDescriptor.